Don't show full schemas on error, check all files
Don't leave the validation errors uncaught (laziness). Instead, catch
and record each one. Then, if there are any, show them in the same
order as the requested files. Include only the filename, path, and
message for now.

Additional tuning of the output or verbosity options are possible for
the future.