Begin setting up test rigging
tox.ini with envs for testing and linting, coverage reporting,
twine-check, and release publishing. Makefile wrapper to handle git
tagging + release publish.

Some initial tests on the cached download tool. Refactor of the tool
to be more testable.

Switch to a src/ layout so that components can be broken apart in a
way that is consistent with testing.