Use urllib and pathlib to handle schema inputs
When handling an input, use urllib.parse.urlparse to get the scheme,
treating no-scheme as a file-like scheme. The scheme is translated
into a "mode" enum, which determines how the SchemaLoader is behaving.

A new error is added to handle unsupported schemes (e.g. ftp://).

Pathlib is used to read and resolve paths, replacing use of
os.path.expanduser.

Tests are adjusted minimally as necessary.

Co-Authored-By: Dan Kolepp <dan@kolepp.dev>