Refactor loader to split local vs remote cases
The schema loader now uses dedicated objects for the local vs remote
schema reading case. This simplifies the initialization logic because
we can focus on parsing URL-like strings and deciding on which
"reader" to use, without trying to also track attributes which will be
needed for file reading in a combined class.

The resulting simplifications make it easier to make path resolution
consistent and testing-friendly.