Pass base_uri to RefResolver, for local schemas
Rather than relying upon the implicit construction of a RefResolver,
pass the input schema path to a RefResolver which we make for any
schema. In the case of a local file, this is the URI-ized path, and in
the case of a remote schema it's the URI used to fetch the schema.

No extra parsing steps are performed on URIs (e.g. removal of
fragments), as these may really be parts of the expected base URI. The
exact meaning in such scenarios is not 100% clear.

Remove a test which is problematic because it gets messed up by
pathlib internals being hard to mock.