Refactor to avoid passing info into schema loader
The schema loader should be solely concerned with reading data from
the filesystem or remote location provided, not with various options
to jsonschema behaviors. For this reason, move the validator building
code from the SchemaLoader object into a standalone helper method, and
remove anything relevant to that from the SchemaLoader (namely, the
format_enabled argument and attribute).