Add vendorized versions of schemas as failover
To support a workflow for releasing new versions of check-jsonschema
with vendorized versions of the schemastore schemas, a script for
updating the vendored schemas is included. It parses the hook config
to pick up on schemas.

The schemas are listed in MANIFEST.in to ensure that they're included
in the dist, and loaded by importlib.resources (or the backport
thereof for py3.6).

The mechanism is controlled via a CLI flag,
`--failover-builtin-schema`. The options for this flag are not yet
documented, but they are just the names of the schemas in the vendor
dir. The usage in the hooks is updated to use this.

resolves #21