Refactor to avoid continuously passing same args
The decomposition of check-jsonschema into a checker and loaders, and
the schema loader further into readers and a cache component, all gets
"undone" to some degree by passing all CLI args through the checker.

To simplify, the parsing is still done by a dedicated function, but in
`main` we use the parsed arguments directly to build several objects
(with appropriate helper funcs). As a result, the checker no longer
sees arguments which don't impact its own behaviors, like
`--no-cache`. These are passed more directly to the relevant
components. As a result, the builtin schema loader is more easily
constructed.