Add finer control for regex checking
`--format-regex` can control the way in which regex format strings are
checked. The default format is now a customized check that treats
special groups which aren't known in python as indicative of another
language's specializations. These are therefore skipped and treated as
"okay".

resolves #20

This also refactors a bit of the way that the format checker gets
built, to support this level of customization.