Add `--data-transform` feature
This feature is built to allow us to define transforms which are
applied to any instancefile after loading, and before validation.

In the case of azure-pipelines, we have a reference implementation in
the azure-pipelines-language-server to look at. The `transforms`
subpackage contains a reimplementation of that same "parser" in
python. For now, only named transforms are supported, and the only
implemented transform is "azure-pipelines".

Update the hook for pre-commit to use
`--data-transform azure-pipelines`.

Include a test pulled from the pipelines-language-server (with
license) to confirm the result.