Remove `--failover-builtin-schema`
This option made more sense prior to the existence of a toolchain to
ensure that the vendored schemas are kept up-to-date. It was
potentially useful to try to use the latest schemastore schema and
failover to the builtin vendored copy.

However, now that there is an automated daily job to update the
vendored schemas, it's much less likely that there's significant drift
between the vendored schema and the schemastore copy. In the event
that there *is* such drift (e.g. a period during which I cannot
perform a release), users can still download the latest schema and
apply it themselves.

Because the failover could lead to inconsistent hook behaviors, it was
removed from the pre-commit hook definitions. Now it is being removed
from the application altogether.