Add support for JSON5 files
Expand the renovate hook config to cover json5 files, and add
conditional support for the json5 format whenever 'json5' or 'pyjson5'
is installed.

The instance loader component is moved from being a standalone module
to being a subpackage, so that json5 support can be (mostly) isolated
into a standalone module.

For now, absent support for json5 in `identify`, there's a special
check for `.json5` as the file extension, to pick up support.

Also, fix the behavior when a BadFileTypeError is raised to use the
shortened trace (not full error output), so that the new "nice" error
is rendered in a more friendly way.