Add support or `--check-metaschema`
This option is mutex with `--schemafile` and `--builtin-schema`, and
indicates that the instancefiles are, themselves, schemas. The correct
validator is selected and built from the relevant metaschema specified
by `"$schema"`, and then applied to each file.

As performance is not yet known to be critical, a new validator is
built per instance when this option is used, ensuring that each
instance is validated under the correct metaschema when they are not
all uniform.