Add check-metaschema hook and use it in tests
Rather than dedicated testing for `metaschemas/...`, the metaschema
hook can now be tested under the hook example files.

The configuration is basically the same as any other hook. As there
are schema-authors interested in writing YAML schemas, the YAML format
is supported for this purpose.

Add some YAML schema files for metaschema tests