Ensure tests cover additional urlinfo schemes
Make sure positive and negative test cases pass for 'file://' schemas.
Add a negative test case for a non-file, non-http(s) scheme, 'ftp://'
is used in the test as a "plausible" example.

In order for the test to pass, explicit handling needs to be
introduced for the UnsupportedUrlScheme error case, and the schema
reader object needs to be initialized lazily.