Set utf-8 encoding for instance files
This can be made configurable or can be detected from the BOM or
chardet in the future. For now, this is right for most users and we
can wait for complaints.

Add a test case for a file with unicode chars. This failed on windows
and was fixed by the use of UTF-8.