Rename CommandState and move instancefile args
'CommandState' is now 'ParseResult' to draw a clearer boundary between
parsing -- in which data is assembled and organized for use -- and
execution -- when the data is used to drive the application.
In concert with this, move build_schema_loader off of the ParseResult
and make it a standalone helper.

Instancefile loading parameters are now attached to the ParseResult.