Simplify CLI parsing code with a more explicit API
Rather than using the click context for the parsed arguments object,
just make creating it and populating it the responsibility of the main
CLI entry point. Then pass it down to the execution layer explicitly,
rather than letting it be found implicitly on the click.Context .

This means that each parsed CLI opt must be defined on the ParseResult
class, but it also leads to a simpler internal API which should enable
better testing.