Add `--quiet` and refine output behaviors
If `-q` is given on text output, no output is produced.
If `-q` is given on JSON output, only the `"status"` field is present
in output.

Reporters require a `verbosity` argument (no default) and the default
verbosity has been moved from 0 to 1 (so that 0 indicates a single
--quiet).

Reporter tests use several verbosity levels.