Improve format support, especially for TOML
Improved format support includes better specifying which checkers to
use based on the JSON Schema draft/version in use. Without this,
attempting to format-validate TOML data which parses as datetime or
time types is difficult, as the draft03 time format checker was
used by default.

In order to better exercise this behavior, `jsonschema[format]` is
added to the tox matrix and run with and without the TOML tests.