Update vendor-schemas to run pre-commit fixers
After a schema is downloaded in the vendor-schemas job, it could be
fixed by pre-commit.ci automatically. The result is an extra push and
noise on the auto-update PR. It's better if the CI job automatically
runs pre-commit on the results (or if the vendored schemas were
ignored under pre-commit; nearly equivalent).
This also ensures that if the only change is that the downloaded
schema reverts the autofixing done under pre-commit, nothing will be
recorded as changed.

Update the gitlab schema by running `make vendor-schemas`.