Implement gitlab `!reference` support
This adds a new --data-transform to the CLI named `gitlab-ci` which
enables `!reference` tag support in the YAML parser.

In order to support this, the transforms have been converted from
callables to objects containing callables. The azure transform has an
"on_data" call (the main data transformer) and the gitlab transform
has an "on_init" call (the YAML parser tweak, to enable the tag).

Transform.on_data is used to modify the document after loading.
Transform.on_init is called when the instance loader is initialized.
Maybe there's a better time for this to happen, but this works today.

A positive test case is included.