Alter the way that the gitlab-ci transform applies
Convert Transform objects from being containers (really, named
tuples) of callables to being a more complete type. YAML instance
customization is a method, so it can be overriden by subclassing, and
`__call__` is defined to check if there's an "on_data" call and invoke
it if so.

The other major change is who/what applies the Transform. This is now
passed down to the data load function, not applied in the
InstanceLoader. This makes it easy for the yaml loading func to access
the Transform (and therefore customize the per-format behavior).

The final "trick" of this change is to cache the modified YAML
implementation based on the Transform used. This ensures that the
Transform YAML tweak isn't applied and re-applied per file, but only
once.