Refactor Transform->instance-load composition
Broadly speaking, the goal of this commit is to undo the requirement
that each loader supports passing in the Transform when this only has
special meaning in the YAML case.

To accomplish this, each InstanceLoader contains a set of parsers to
use (rather than this being a global dict). The global dict is still
present in the form of a set of defaults, but it is copied when the
set of parsers is being setup. This allows the
"modify_yaml_implementation" hook to be called during InstanceLoader
setup to create a valid "yaml load" function to use.