Add unit tests for gitlab transform, fix minor bug
Add unit tests and add the note to the changelog.
The tests expose a minor issue in the way that ruamel.yaml.YAML
instances share state via class attributes on constructor classes.

Although the situation is not detrimental to check-jsonschema today,
it's a surprising and undesirable side-effect. It also plays havoc on
the testability of the code, since custom constructors are
"accidentally" shared. To workaround it, use a dynamically generated
Constructor subclass per YAML instance.