Refactor multiple parser support to reuse
InstanceLoader and SchemaLoader now share the ParserSet impelmentation
of multiple parsers under `identify`-driven dispatch. With some minor
adjustments to support the schema case, this now supports both and
reduces the possibility of fixes for one case not applying the other.
e.g. the datetime handling in the instance YAML loader is now applied
to the schema loader.

This allows support for JSON5 to be added to the SchemaLoader fairly
trivially, with an added check in the RefResolver similar to the
existing YAML check.

resolves #102