Do not crash on instance parse errors
Pass instancefile parsing errors down to the Reporter and produce
appropriate output to communicate a mixture of parsing and validation
errors.

If a parse error is encountered, the run will always exit(1)
(failure), but validation will still run on successfully parsed
instances and output any validation errors of interest.

In order to support this change in behavior, the Reporter now consumes
a CheckResult object, rather than the raw validation error map. This
allows the reporter to handle the multiple "streams" of error
information through a clean interface. It also means that there is now
an object responsible for the filename->error_list mapping interface.