Make azure-pipelines transform more permissive
Cases which were previously rejected are now dropped from the data
when a pipeline is being transformed.
This can result in certain data transforming to an empty object, which
*can* pass validation. What it can't do is provide a required or
expected key based on the value which is produced by the parameter
expansion.

For example, the following user-provided example:

        env:
          ${{ each pair in parameters.env }}:
            ${{ pair.key }}: ${{ pair.value }}

should now effectively "expand" to

        env: {}