Setup json validation for remote schemafiles (#184)
First, add support to the cachedownloader for validation.
A validation callback method can be passed to the cachedownloader on
init, and will be used to test data and potentially retry before
returning a result. The basic premise is that the validator should be
a function which can raise a ValueError or a subclass thereof if the
data downloaded doesn't meet our expectations (e.g. a partial download
or empty data returned as a 200).

Then apply this to the schema reader by passing `json.loads` as a
validation callback. This will raise a JSONDecodeError on malformed
data, triggering the retry.

Add the appropriate note to the changelog.
resolves #183