Implement rudimentary `--fill-defaults` flag
Docs list this among the various flags, three acceptance tests
exercise the range of expected/desired functionality, and the
changelog lists the new feature.

No explicit handling is added for polymophic schemas, incorrect
`"default"` values, or other corner cases, but the docs include a
warning that `"default"` can be ambiguous or even ill-defined.

Although the behavior, inherited from `jsonschema`, is deterministic
for how `"default"` under a polymorphic schema is handled, it is
documented explicitly as undefined.