Improve testing of missing toml parser
Add a new constant (in a dummy module) which can be used to explicitly
disable TOML parsing, solely for the purpose of testing. Comments on
the variable explain why it is necessary.

Also cleanup tox config a little. Remove the depends listing in favor
of ordering environments in the env_list in a sensible way. This
allows for parallel invocation only when the coverage environments are
absent, but this is still possible. For example,
`tox r -e cov_clean; tox p -m ci; tox r -e cov` would imitate CI
usage.