Add a check for tests which are always skipped
This takes the form of a new script which crawls pytest XML reports
and collates them into a single aggregate. It checks for tests which
are skipped or missing in all of the reports.

The aggregator can be run along with a suite of tox environments via
`make collatd-test-report`, and a new CI job runs this in a build.