Add a check for tests which are always skipped (#240)
* Add a check for tests which are always skipped

This takes the form of a new script which crawls pytest XML reports
and collates them into a single aggregate. It checks for tests which
are skipped or missing in all of the reports.

The aggregator can be run along with a suite of tox environments via
`make collatd-test-report`, and a new CI job runs this in a build.

* Don't run tests twice in CI

As a first draft, tests ran twice (one for the "test" and once for the
skipped test collator). Switch to upload/download of junitxml report
data to pass the reports from the matrix to the collator.