Catch catalog drift vs generated hooks
Add generated hook comments to help contributors, instructing them to
modify the catalog.py + a GHA job to check for catalog discrepancies.

Contributors easily modify `.pre-commit-hooks.yaml` without modifying
the catalog file, which puts us in the state that changes can easily
be accidentally reverted. The job enforces that if regenerating the
hooks produces a change, the PR fails in CI, and the comments help
contributors avoid failing CI.