Fix the handling of bad Last-Modified headers (#277)
Originally found via an OverflowError on some Windows usage where the
default Last-Modified header was used, there are several issues which
were not handled correctly by the downloader.
1. OverflowError: a last mod date outside of `mktime`'s supported
   range causes this
2. ValueError: a malformed header causes `strptime` to fail
3. LookupError/KeyError: the header is missing

(3) was previously handled by falling back to a default time which was
then strptime parsed. This is now caught as a LookupError and handled
with the fixed default of `0.0` used for all of these bad-parse cases.
The semantics are the same, but the implementation is simpler and more
uniform.