Introduce remote $ref tests and fix URL join logic
URL joining needs to be done between any relative ref paths and the
base URI. This was not correct, as revealed by a new testcase.
After fixing, this revealed that the local file loading was somewhat
indirect in a way that it no longer needs to be, so local file
handling in the ref resolver has been refactored as well.