Improve internal terminology for retrieval URI
The retrieval URI was internally called the 'ref base' or 'schema
URI', both of which suggest that it is definitively the base URI used
for reference resolution. This makes it confusing that adjusting this
URI might not change the base URI used.

Instead, call it properly 'retrieval URI'.

Base URI is newly introduced inside of the retrieve callback, after
receiving both the `$id` (if any) and the retrieval URI.