Convert build pipelines to use two jobs
This avoids exposing the GitHub token (`id-token: write`) used for
trusted publishing to the build process (`build` and the underlying
backend), improving security via isolation.

Also, add `twine check` to the build+publish pipeline, per the
recommendations of various publishing guides, to catch malformed
metadata.

Big thanks to @webknjaz for spotting this improvement and for
providing tools, docs, and guidance for publishing!