Fix a bug in GHA due to unclear docs
One thing is documenting the event payloads, another is documenting
the event types, etc.

The end result is that this is a highly non-obvious string which we'll
have to guess until we get it right. And because it's a release-only
activity, the best way would be to take some dedicated testing time.
Since it's non-urgent, just fiddle with it here for now and revisit
later.