Add interface for passing custom validators
This adds the necessary pieces for passing a custom validator class to
check-jsonschema. `--fill-defaults` help text contains a hint that it
may conflict (depending on the validator), to help warn of the
potential for surprising interplay between a custom validator class
and optional check-jsonschema behaviors.

This commit only adds the interface pieces for passing the validator
class to the CLI in a well-specified format, and handles the
import/load of that value. No implementation in terms of threading the
resulting value down to the validator(s) is included yet.