Apply click-type-test to check CLI annotations
Using click-type-test, we can confirm that the parameter annotations
are correct.

Most of the annotations line up. The only things which needed special
attention were `Literal` deductions from `click.Choice` parameters. A
couple of the types are easily tripped up by some of the dynamism at
play, so these are solved with `overrides` in the typing test.

There's also an adjustment to fix a BinaryIO/IO[bytes] discrepancy,
driven mostly to become consistent with the click type annotations.