Fix collation script to handle module-level skips
This is meant to detect always-skipped tests by checking for skips
which never appear as non-skips.

However, a module-level skip produces a skip for the module which
doesn't *also* get emitted as a success on the "positive" runs. Since
module-level skips have an empty `classname`, use that to detect them
and treat them slightly differently (by counting them and comparing
against the number of files).