Make TOML parsing always available
This makes the dependency on `tomli` for older Pythons non-optional,
ensuring that `.toml` files are always supported.

Most of the code removals have to do with testing and supporting the
state with no TOML parser available. There's no longer an `ENABLED`
bool control, and tests no longer try to use that to check the
behavior in the absence of a parser.