Expand caching to cache remote refs
Several refinements are needed in the CacheDownloader to support this.
Primarily, support for sibling directories to the `downloads` dir, in
the cache dir. This allows the ref resolver to pass in `"refs"` as a
directory name.

As a related change in the course of this work, HTTP retries are
expanded in scope to also cover connection errors and timeouts.

Additionally, `disable_cache` gets passed down from the CLI through to
the ref resolution layer.

Tests are enhanced to better explore CacheDownloader behaviors, but
not to test the usage in ref resolution.