Add tests for remote YAML ref loading and caching
With any non-`.json` suffix, e.g., `.yaml`, we need the cache to
preserve extensions in order to be certain that the cache loader will
choose the same parser when loading the cached ref as the one it used
when the ref was remote. The `.yaml` ref + cache population trickery
allows us to test this codepath in an acceptance test.