Refactor cache-related fixtures
Add new fixtures which expose nicer interfaces for interacting with
the download caches. These are split into two variants -- one for
`refs/` and one for `downloads/` -- for the relevant functions. The
new fixtures cover getting the relevant cache dir as a path, getting
the expected cache location for a URI as a path (without creating it),
and injecting some specific file contents into said path, including
creation of the parent dir structure.

These new fixtures are then applied in as many locations as possible
for a small but valuable reduction in test complexity and tight
coupling between specific tests and the implementation.