Instrument `regress`-backed `pattern` evaluation
When an alternate RegexImplementation is selected, it will be passed
down to the code which builds a Validator. The Validator is extended
with the keyword validator provided by the RegexImplementation.

Because this uses the `extend()` interface, a test which subclassed a
validator broke -- this is documented in `jsonschema` as unsupported
usage, so the test simply had to be updated to use supported
interfaces.