Add a nonunicode regex mode
This is a variant of the regress engine usage in which the `u` flag is
not set. It is needed for usages which rely on unicode escapes *not*
being interpreted. In particular, the azure pipelines schema won't
pass metaschema validation (let alone apply correctly) if `pattern`
interpretation uses unicode-mode regexes.