Use regex variant when checking schema validity
Rather than relying directly on `validator_cls.check_schema`,
reimplement this functionality in-tree here to allow for
customizations to the validator class before it is run on the user's
schema.

A new test uses an invalid regex under the regress unicode engine
which is valid in the python engine to ensure that consistent checking
is applied.

Manual testing revealed that the `_fail()` message production for
SchemaError was showing error information twice, which is fixed
without a new test to guarantee the new behavior.