Always enable metaschema format checking
In the implementation prior to this change, passing
`--disable-formats` would impact not only the "actual" schema
validator, but also the validator built to evaluate the schema against
its metaschema. As a result, `--disable-formats *` and similar would
enable schemas to run which previously should have been caught as
invalid.

Furthermore, the customized format checker which had extensions for
date and time evaluation added was used, and any other customizations
to format checking would implicitly be shared with the metaschema
check.

To resolve, refactor format checker building to allow it to be used
more directly for the metaschema check, and add test cases to confirm
that a bad regex in a `pattern` is always rejected, even when
`--disable-formats regex` or similar is used.